Add some integration tests, and further flesh out the functionality
The integration tests exposed the inconvenience of the API. For now,
we add 'Java' versions of the ItemCodec helpers. This means it is possible
to encode and decode data directly from the (Java) AWS DynamoDB client,
without needing to convert to and from Java types.
This probably wants more thought in the future.

To support the integration tests, we add the sbt-dynamodb plugin. This
plugin is no longer maintained, so this is a temporary solution.